+++ Deshaker/DummyFrame.cpp
@@ -32,7 +32,7 @@ int64_t DummyFrame::createPyramid(int64_t frameIndex, AffineDataFloat trf, bool 
 	return 0;
 }
 
-void DummyFrame::adjustPyramid(int64_t frameIndex, double gamma) {}
+void DummyFrame::adjustPyramid(int64_t frameIndex, float gamma) {}
 
 Matf DummyFrame::getTransformedOutput() const { 
 	return {};
